+++ src/app/Http/Controllers/Admin/VoterController.php
@@ -3,21 +3,37 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Http\Resources\ElectionResource;
-use App\Models\Election;
+use App\Imports\UserImport;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use Maatwebsite\Excel\Facades\Excel;
+use Maatwebsite\Excel\Importer;
 
 class VoterController extends Controller
 {
-    public function show(Election $election)
+    private $importer;
+
+    public function __construct(Importer $importer)
     {
-        $voter = User::all();
+        $this->importer = $importer;
+    }
+
+    public function index()
+    {
+        $users = User::get();
 
         return Inertia::render('Admin/Voters/index', [
-            'election' => new ElectionResource($election),
-            'voters' => $voter,
+            'users' => $users,
         ]);
+    }
+
+    public function import(Request $request)
+    {
+        $file = $request->file('file');
+        $import = new UserImport();
+        Excel::import($import, $file, null, \Maatwebsite\Excel\Excel::CSV);
+
+        return to_route('admin.voters.index');
     }
 }
