--- src/app/Http/Requests/Admin/ImportFormRequest.php
@@ -1,32 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Admin;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class ImportFormRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
-     */
-    public function rules(): array
-    {
-        return [
-            'upfile' => [
-                'required',
-                'mimetypes:application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-                'mimes:csv',
-            ]
-        ];
-    }
-}
