+++ src/resources/js/Layouts/Navbar/MainMenu.jsx
@@ -7,7 +7,7 @@ function MainMenu ({ className }) {
         <div className={className}>
             <MenuItem text="概要" link="overview" icon="overview" />
             <MenuItem text="投票者" link="voters" icon="voters" href={route('admin.voters.index')} />
-            <MenuItem text="立候補者" link="candidates" icon="candidates" href={route('admin.candidates')} />
+            <MenuItem text="立候補者" link="candidates" icon="candidates" href={route('admin.candidates.index')} />
             <MenuItem text="結果" link="result" icon="result" />
         </div>
     );
