+++ src/resources/js/Pages/Admin/Dashboard.jsx
@@ -4,7 +4,7 @@ import { Head, Link } from '@inertiajs/react';
 import ElectionCard from './ElectionCard';
 import PageHeader from '@/Components/dashboard/PageHeader';
 
-export default function Dashboard({auth, elections}) {
+export default function Dashboard({ auth, elections }) {
     return (
         <AuthenticatedLayout
             auth={auth}
@@ -25,7 +25,6 @@ export default function Dashboard({auth, elections}) {
                                                 electionName={item.election_name}
                                                 electionStartDate={item.start_date}
                                                 electionEndDate={item.end_date}
-                                            
                                             />
                                         </Link>
                                     </div>
