--- src/resources/js/Pages/Admin/ElectionCreate.jsx
@@ -1,10 +1,0 @@
-export default function ElectionCreate({ auth }) {
-    return(   
-    <AuthenticatedLayout
-        user={auth.user}
-        // header={<h2 className="font-semibold text-xl text-gray-800 leading-tight">Dashboard</h2>}
-    >
-       
-    </AuthenticatedLayout>
-    )
-}
