+++ src/resources/js/Pages/Admin/Overview/index.jsx
@@ -1,9 +1,14 @@
+import LaunchButton from '@/Components/overview/LaunchButton'
 import ElectionLayout from '@/Layouts/ElectionLayout'
 import React from 'react'
 
 const Overview = () => {
     return (
-        <div>概要</div>
+        <div>
+            概要
+
+            <LaunchButton handleClick={null} />
+        </div>
     )
 }
