--- src/resources/js/Pages/Admin/election/Voters.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Voters() {
-    return (
-        <div>Voters</div>
-    )
-}
-
-export default Voters
